+++ src/TagLib/IFD/Entries/SubIFDEntry.cs
@@ -29,13 +29,13 @@ namespace TagLib.IFD.Entries
 	/// </summary>
 	public class SubIFDEntry : IFDEntry
 	{
-		public uint Tag { get; private set; }
+		public ushort Tag { get; private set; }
 		public uint Type { get; private set; }
 		public uint Count { get; private set; }
 
 		public IFDStructure Structure { get; private set; }
 
-		public SubIFDEntry (uint tag, uint type, uint count, IFDStructure structure)
+		public SubIFDEntry (ushort tag, uint type, uint count, IFDStructure structure)
 		{
 			Tag = tag;
 			Type = type;
